Add class in their own file
