+++ base.rb
@@ -10,16 +10,3 @@ class Decorator < Nameable
     @nameable.correct_name
   end
 end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    @nameable.correct_name.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    trimmed_name = @nameable.correct_name
-    trimmed_name.length > 10 ? trimmed_name[0..9] : trimmed_name
-  end
-end
