+++ capitalize.rb
@@ -1,0 +1,7 @@
+require_relative 'base'
+
+class CapitalizeDecorator < Decorator
+  def correct_name
+    @nameable.correct_name.capitalize
+  end
+end
